requirements.txt
Django==1.3.1
Fabric==1.3.3
South==0.7.3
Werkzeug==0.8.3
coverage==3.5.1
django-autoslug-field==0.2.3
django-debug-toolbar==0.9.4
-e git+https://github.com/howiworkdaily/django-faq.git#egg=django-faq
-e git+https://github.com/django-extensions/django-extensions.git#egg=django-extensions
django-nose==0.1.3
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=django-registration
django-taggit==0.9.3
gunicorn==0.13.4
ipython==0.10.1
nose==1.1.2
pycrypto==2.4.1
simplejson==2.3.0
ssh==1.7.11
wsgiref==0.1.2
Whoosh==2.3.2
-e git+https://github.com/toastdriven/django-haystack.git@4fb267623b58c5581be96a9a9504ca10a72eb0d8#egg=django_haystack-dev
